backend/requirements.txt
asgiref==3.3.4
boto3==1.17.59
botocore==1.20.59
dj-database-url==0.5.0
Django==3.2
django-environ==0.4.5
django-heroku==0.0.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
jmespath==0.10.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
